+++ src/api/routes/v1/payment-gateway.route.ts
@@ -1,0 +1,13 @@
+import bodyParser from 'body-parser'
+import {Router} from 'express'
+import {
+  paymentGatewayWebhookMiddleware,
+  paymentGatewayWebhookRoute
+} from '../../controllers/payment-gateway.controller'
+const app = Router({mergeParams: true})
+app.post('/webhook',
+  // paymentGatewayWebhookMiddleware, 
+  paymentGatewayWebhookRoute
+)
+
+export default app
