--- src/api/tests/unit-testing/user/sign-in.ts
@@ -1,7 +1,0 @@
-// import 
-
-// export default () => {
-//   describe('@Sign-In Entity', () => {
-
-//   })
-// }
